create global-ui file for all main components utils functions
